scripts/requirements.txt
duckdb>=0.10.0
google-genai>=0.8.0
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0 


